requirements.txt
boto3==1.35.5
botocore==1.35.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
dbt-redshift==1.8.1
dbt-core==1.8.6